+++ AppStore/Controller/OtherPeopleTableViewController.swift
@@ -7,21 +7,17 @@
 //
 
 import UIKit
-import NVActivityIndicatorView
 
 class OtherPeopleTableViewController: UITableViewController {
     
     var userId = ""
     var reviewArray: [Review] = []
-    var itemArray: [Item] = []
     var user: User!
-    var activityIndicator: NVActivityIndicatorView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
         tableView.tableFooterView = UIView()
-        activityIndicator = NVActivityIndicatorView(frame: CGRect(x: self.view.frame.width / 2 - 30, y: self.view.frame.height / 2 - 30, width: 60.0, height: 60.0), type: .ballClipRotatePulse, color: UIColor(named: "original yellow"), padding: nil)
         loadUser()
     }
     
@@ -39,22 +35,16 @@ class OtherPeopleTableViewController: UITableViewController {
     
     private func loadReview() {
         
-        showLoadingIndicator()
         downloadReview(self.user.reviewId) { (allReview) in
             self.reviewArray = allReview
-            self.title = "レビュー数: \(self.reviewArray.count)"
-            
-            downloadItems(self.user.itemId) { (allItem) in
-                self.itemArray = allItem
-            }
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
+            self.title = "レビュー数 \(self.reviewArray.count)"
+            DispatchQueue.main.async {
                 self.tableView.reloadData()
-                self.hideLoadingIndicator()
             }
         }
     }
     
-    // MARK: - Table view data source
+    // MARK: Table view data source
     
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         
@@ -62,10 +52,8 @@ class OtherPeopleTableViewController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        let indexNumber = indexPath.row
-        
-        if indexNumber == 0 {
+                
+        if indexPath.row == 0 {
             
             let cell = tableView.dequeueReusableCell(withIdentifier: "Cell") as! OtherPeopleTableViewCell
             
@@ -78,49 +66,11 @@ class OtherPeopleTableViewController: UITableViewController {
         return cell2
     }
     
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
-        let indexNumber = indexPath.row
-        if indexNumber > 0 {
-            
-            tableView.deselectRow(at: indexPath, animated: true)
-            showItemView(itemArray[indexPath.row - 1])
-        }
-    }
-    
-    //MARK: Navigation
-    
-    private func showItemView(_ item: Item) {
-        
-        let detailVC = UIStoryboard.init(name: "Main", bundle: nil).instantiateViewController(identifier: "DetailVC") as! DetailTableViewController
-        detailVC.item = item
-        
-        self.navigationController?.pushViewController(detailVC, animated: true)
-    }
-    
     //MARK: IBAction
     
     @IBAction func backButtonPressed(_ sender: Any) {
         
         self.navigationController?.popViewController(animated: true)
-    }
-    
-    //MARK: Activity Indicator
-    
-    private func showLoadingIndicator() {
-        
-        if activityIndicator != nil {
-            self.view.addSubview(activityIndicator!)
-            activityIndicator!.startAnimating()
-        }
-    }
-    
-    private func hideLoadingIndicator() {
-        
-        if activityIndicator != nil {
-            activityIndicator!.removeFromSuperview()
-            activityIndicator!.stopAnimating()
-        }
     }
     
 }
