+++ AppStore/GoogleService-Info.plist
@@ -3,9 +3,9 @@
 <plist version="1.0">
 <dict>
 	<key>CLIENT_ID</key>
-	<string>747258727126-p8gab2fb6ns1du9b4c8f4oh7i39ids4b.apps.googleusercontent.com</string>
+	<string>747258727126-mvmj2gbg3ujkfmurbfnt5uta2gvvcucl.apps.googleusercontent.com</string>
 	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.747258727126-p8gab2fb6ns1du9b4c8f4oh7i39ids4b</string>
+	<string>com.googleusercontent.apps.747258727126-mvmj2gbg3ujkfmurbfnt5uta2gvvcucl</string>
 	<key>API_KEY</key>
 	<string>AIzaSyADc3AweBuW3cZ2CcqUubWC7_9F0j5Skjw</string>
 	<key>GCM_SENDER_ID</key>
@@ -13,7 +13,7 @@
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
-	<string>yncompany.AppStore</string>
+	<string>yncompany.StoreApp</string>
 	<key>PROJECT_ID</key>
 	<string>appstore-a772d</string>
 	<key>STORAGE_BUCKET</key>
@@ -29,7 +29,7 @@
 	<key>IS_SIGNIN_ENABLED</key>
 	<true></true>
 	<key>GOOGLE_APP_ID</key>
-	<string>1:747258727126:ios:d0809e08f4060261d1d777</string>
+	<string>1:747258727126:ios:5adf45f10386792ad1d777</string>
 	<key>DATABASE_URL</key>
 	<string>https://appstore-a772d.firebaseio.com</string>
 </dict>
