+++ StoreApp.xcodeproj/xcuserdata/nakamotoyuuji.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,12 @@
 		<key>AppStore.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>37</integer>
+			<integer>29</integer>
+		</dict>
+		<key>StoreApp.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>29</integer>
 		</dict>
 	</dict>
 </dict>
